examples/scripts/monitoring/monitor_yubikey_network.py
#!/usr/bin/env python3

"""
YubiKey Network Monitoring Script
This script implements monitoring, metrics collection, and alerting for the
YubiKey-Based Autonomous Sub-Key Network.
"""

import os
import sys
import time
import logging
import argparse
from typing import Dict, List, Optional
from datetime import datetime, timedelta
import yaml
import prometheus_client
from prometheus_client import Counter, Gauge, Histogram, start_http_server
import requests
import smtplib
from email.message import EmailMessage

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger(__name__)

class YubiKeyMetrics:
    """Prometheus metrics for YubiKey operations and health."""
    
    def __init__(self):
        # YubiKey Operations
        self.operations = Counter(
            'yubikey_operations_total',
            'Total number of YubiKey operations',
            ['operation_type', 'status']
        )
        
        # YubiKey Health
        self.health = Gauge(
            'yubikey_health_status',
            'YubiKey health status (1 = healthy, 0 = unhealthy)',
            ['yubikey_id']
        )
        
        # Operation Latency
        self.latency = Histogram(
            'yubikey_operation_duration_seconds',
            'Time spent processing YubiKey operations',
            ['operation_type']
        )
        
        # Authentication Metrics
        self.auth_failures = Counter(
            'auth_failures_total',
            'Authentication failures',
            ['reason']
        )
        
        # Key Management
        self.key_rotations = Counter(
            'key_rotations_total',
            'Key rotation events',
            ['key_type']
        )
        
        # Active Sessions
        self.active_sessions = Gauge(
            'active_sessions',
            'Number of active sessions'
        )

class AlertManager:
    """Manages alerting for the YubiKey network."""
    
    def __init__(self, config: Dict):
        self.config = config
        self.load_alert_rules()
        
    def load_alert_rules(self):
        """Load alert rules from configuration."""
        try:
            with open(self.config['alert_rules_path'], 'r') as f:
                self.rules = yaml.safe_load(f)
        except Exception as e:
            logger.error(f"Failed to load alert rules: {e}")
            self.rules = {}
            
    def check_alerts(self, metrics: YubiKeyMetrics):
        """Check metrics against alert rules."""
        for rule in self.rules.get('rules', []):
            try:
                if self.evaluate_rule(rule, metrics):
                    self.send_alert(rule)
            except Exception as e:
                logger.error(f"Failed to evaluate rule {rule['name']}: {e}")
                
    def evaluate_rule(self, rule: Dict, metrics: YubiKeyMetrics) -> bool:
        """Evaluate a single alert rule."""
        try:
            # Example rule evaluation
            if rule['type'] == 'threshold':
                value = self.get_metric_value(metrics, rule['metric'])
                threshold = float(rule['threshold'])
                
                if rule['comparison'] == 'above':
                    return value > threshold
                elif rule['comparison'] == 'below':
                    return value < threshold
                    
            return False
        except Exception as e:
            logger.error(f"Rule evaluation failed: {e}")
            return False
            
    def get_metric_value(self, metrics: YubiKeyMetrics, metric_name: str) -> float:
        """Get current value of a metric."""
        try:
            # Example metric value retrieval
            if metric_name == 'auth_failures_rate':
                return metrics.auth_failures._value.get()
            elif metric_name == 'yubikey_health':
                return metrics.health._value.get()
            return 0.0
        except Exception as e:
            logger.error(f"Failed to get metric value: {e}")
            return 0.0
            
    def send_alert(self, rule: Dict):
        """Send alert notification."""
        try:
            if 'email' in self.config['alert_channels']:
                self.send_email_alert(rule)
            if 'slack' in self.config['alert_channels']:
                self.send_slack_alert(rule)
        except Exception as e:
            logger.error(f"Failed to send alert: {e}")
            
    def send_email_alert(self, rule: Dict):
        """Send alert via email."""
        try:
            msg = EmailMessage()
            msg.set_content(f"Alert: {rule['name']}\nSeverity: {rule['severity']}\nDescription: {rule['description']}")
            
            msg['Subject'] = f"YubiKey Network Alert: {rule['name']}"
            msg['From'] = self.config['email']['from']
            msg['To'] = self.config['email']['to']
            
            with smtplib.SMTP(self.config['email']['smtp_server']) as server:
                server.send_message(msg)
                
            logger.info(f"Sent email alert for rule: {rule['name']}")
        except Exception as e:
            logger.error(f"Failed to send email alert: {e}")
            
    def send_slack_alert(self, rule: Dict):
        """Send alert via Slack."""
        try:
            payload = {
                'text': f"*YubiKey Network Alert*\n"
                       f"*Rule:* {rule['name']}\n"
                       f"*Severity:* {rule['severity']}\n"
                       f"*Description:* {rule['description']}"
            }
            
            response = requests.post(
                self.config['slack']['webhook_url'],
                json=payload
            )
            response.raise_for_status()
            
            logger.info(f"Sent Slack alert for rule: {rule['name']}")
        except Exception as e:
            logger.error(f"Failed to send Slack alert: {e}")

class NetworkMonitor:
    """Main monitoring class for the YubiKey network."""
    
    def __init__(self, config: Dict):
        self.config = config
        self.metrics = YubiKeyMetrics()
        self.alert_manager = AlertManager(config)
        self.setup_logging()
        
    def setup_logging(self):
        """Configure logging based on configuration."""
        if self.config.get('debug'):
            logger.setLevel(logging.DEBUG)
            fh = logging.FileHandler('monitoring.log')
            fh.setLevel(logging.DEBUG)
            logger.addHandler(fh)
            
    def start_metrics_server(self):
        """Start Prometheus metrics server."""
        try:
            start_http_server(self.config['metrics_port'])
            logger.info(f"Started metrics server on port {self.config['metrics_port']}")
        except Exception as e:
            logger.error(f"Failed to start metrics server: {e}")
            sys.exit(1)
            
    def collect_metrics(self):
        """Collect metrics from the YubiKey network."""
        while True:
            try:
                # Collect YubiKey health metrics
                self.collect_yubikey_health()
                
                # Collect authentication metrics
                self.collect_auth_metrics()
                
                # Collect session metrics
                self.collect_session_metrics()
                
                # Check for alerts
                self.alert_manager.check_alerts(self.metrics)
                
                # Wait for next collection interval
                time.sleep(self.config['collection_interval'])
                
            except Exception as e:
                logger.error(f"Metrics collection failed: {e}")
                time.sleep(10)  # Wait before retry
                
    def collect_yubikey_health(self):
        """Collect YubiKey health metrics."""
        try:
            # Example health check implementation
            for yubikey_id in self.config['yubikeys']:
                # Perform health check
                is_healthy = self.check_yubikey_health(yubikey_id)
                
                # Update metric
                self.metrics.health.labels(yubikey_id=yubikey_id).set(
                    1 if is_healthy else 0
                )
        except Exception as e:
            logger.error(f"YubiKey health check failed: {e}")
            
    def check_yubikey_health(self, yubikey_id: str) -> bool:
        """Check health of a specific YubiKey."""
        try:
            # Implement health check logic
            # This is a placeholder - implement actual health check
            return True
        except Exception as e:
            logger.error(f"Health check failed for YubiKey {yubikey_id}: {e}")
            return False
            
    def collect_auth_metrics(self):
        """Collect authentication metrics."""
        try:
            # Example authentication metrics collection
            # This is a placeholder - implement actual metric collection
            pass
        except Exception as e:
            logger.error(f"Auth metrics collection failed: {e}")
            
    def collect_session_metrics(self):
        """Collect session metrics."""
        try:
            # Example session metrics collection
            # This is a placeholder - implement actual metric collection
            pass
        except Exception as e:
            logger.error(f"Session metrics collection failed: {e}")

def main():
    """Main function for the monitoring script."""
    parser = argparse.ArgumentParser(
        description='YubiKey Network Monitoring Tool'
    )
    parser.add_argument(
        '--config',
        required=True,
        help='Path to configuration file'
    )
    parser.add_argument(
        '--debug',
        action='store_true',
        help='Enable debug logging'
    )
    
    args = parser.parse_args()
    
    try:
        # Load configuration
        with open(args.config, 'r') as f:
            config = yaml.safe_load(f)
            
        # Add debug flag to config
        config['debug'] = args.debug
        
        # Create and start monitor
        monitor = NetworkMonitor(config)
        monitor.start_metrics_server()
        monitor.collect_metrics()
        
    except Exception as e:
        logger.error(f"Monitoring failed: {e}")
        sys.exit(1)

if __name__ == '__main__':
    main()