examples/scripts/monitoring/requirements.txt
prometheus_client>=0.16.0
pyyaml>=6.0.0
requests>=2.28.0
typing-extensions>=4.5.0